--- a/Misc/NEWS.d/next/Library/2025-01-10-12-19-42.gh-issue-118761.i4yXe3.rst
+++ b/Misc/NEWS.d/next/Library/2025-01-10-12-19-42.gh-issue-118761.i4yXe3.rst
@@ -1,0 +1,2 @@
+Improve import time of :mod:`secrets` by a factor two. This is achieved by
+importing :mod:`base64` on demand. Patch by Bénédikt Tran.