@AA-Turner 2025-01-12
We should re-evaluate this after `base64` no longer imports `re`.