@picnixz 2025-01-14
Since `base64` will not import `re` anymore, we can perhaps avoid this change as I fear it won't be impactful at all. So I plan to defer this one until #128736 is merged.