@vstinner 2025-01-14
https://github.com/python/cpython/pull/128736 was merged.