@picnixz 2025-01-14
I therefore don't think it's worth the change. Less work for me and less diff with other branches. I'm going to close this PR.