@picnixz 2025-01-14
Thank you Hugo for running the numbers. There is still one module that I had in mind (`pickletools`) since it imports `re` but I haven't checked other modules importing `re`. I'll do that tomorrow if I time, otherwise this w-e.