@vstinner 2025-01-14
I don't think that pickletools import time matters since pickletools is a debug tool.